Interim commit for Requirement: #664
Updates getting a first version of Shreds for the ImageServer project. This commit has a framework for accepting DICOM associations.  It will generate a list of services supported from the database and listen for those services.  It still has to add code for the actual saving of incoming objects to disk and saving information about them in the database.