Requirement: #837 Extensioin Points defined in ClearCanvas.Common don't appear in extension browser
This bug is caused by the fact that ClearCanvas.Common is not technically a plugin, and therefore the PluginManager does not attempt to read its meta-data.  I added a really simple hack to force the plugin manager to read the meta-data of ClearCanvas.Common.  This does not seem to have any adverse effects.

Extension points defined in Common now show up in the extension browser.  However, ClearCanvas.Common does not show up as a plugin.  I believe this is the most appropriate behaviour, given that it is not in fact a plugin.