Update for Bug: #835
Further changes to try and resolve this defect with shutdowns.  Re-worked how we listen for connections after seeing some stalled connections.  We no longer use async I/O when doing this.  Also found an MSDN reference as to detecting when a socket has shut down by checking for an exception when sending a 0 byte buffer.  Incorporated changes to use this new method to detect shutdowns.  This appears to resolve the problem, but further testing is necessary.