Requirement: #772
Partial commit.  This code includes new application settings for if devices should be auto-inserted into the database, and if we should reject associations from devices that are not in the database.  Also, a new Device table was added to the database, stored procedures added to access the table, and updates to the model to support these.  Also made renamed the SopClass table to ServerSopClass so in the model the name does not conflict with the SopClass class in the DICOM library.  Finally, cleaned up some namespaces for parameters to stored procedures in the model.