Commit for Requirement: #678
This commit includes support for image level queries derived from the XML file, and also for determining the negotiated services for an association based on the XML file, as opposed to opening each file to determine its transfer syntax, etc.  The checkin also adds support for a verification SCP extension on the SCP. A bunch of code cleanup is also included as recommended by resharper.