Updates for Bug: #959
Modified SendCFindRequest to have a DicomMessage parameter instead of a DicomAttributeCollection as an input parameter.  This allows setting the priority in the message by the user.  Also cleaned up some of the internal code for sending messages, added routines for sending Cancel messages, and did some other cleanup recommended by resharper.