Requirement: #989 Loaded settings instances not refreshed after modifying settings in RIS application
Modified ApplicationSettingsRegister so that instead of doing funky stuff with the app.config file, it merely calls Reload() on the loaded settings instance of a given type.  This *seems* to work correctly in both the local and enterprise environments, although it is possible that the behaviour in the local environment will not be exactly the same as it was previously.