Commit for Requirement: #995
Commit for using the .NET framework's Asyc pattern for the StorageScu component.  This update also includes adding a base class for SCUs, and the addition of a VerificationScu class for doing verification associations.  Also added some refactoring of the DICOM Library to better handle association shutdowns.  Also includes some general cleanups as recommended by resharper.

This code was submitted on the Forum by Dan Blanchard.