Bug: #1093
Updated the ImageServer code to use the using() construct with the majority of the database connections in use in the code.  Also, updated places where updates are done in the code to use the IUpdateContext instead of the IReadContext (even though with the ADO.NET implementation, updates via stored procedures can be done on the read context).  Also, discovered some problems with the UpdateContext implementation and how it dealt with transactions and cleaned these up.