This commit is for ticket #845: Performance Logging
Code checked in previously doesn't capture the statistics correctly. Only inbound data was reported in the log. The reason was not all transmission "events" are exposed and passed to the recorder.

The code is refactored for this checkin so that the statistic wrapper can listen to network events directly instead of being notified indirectly by the handler (which will require the handler to override the methods in the NetworkBase for it to work properly)

- Changed BaseStatistics namespace to ClearCanvas.Common.Statistics
- Split different statistic classes into separate files.
- Fixed bug in StreamReader reporting the estimate number of bytes in BytesEstimated
- Added fields in AssociationParameters to store inbound and outbound statistics.
- Added events in NetworkBase
- Changed AssociationStatisticsRecorder to listen to events in NetworkBase directly.
- Removed OnAssociationEstablished and OnAssociationRelease in IDicomServerHandler. DicomServerHandler objects should now listen to association events fired by the NetworkBase object instead.
- Fixed sample code for the new interface.