Bug: #816
Fixed problem w/ Changeset [3560] commit.

Changes to resolve this bug were added to deal w/ when the group 2 length was not correct in the meta info.  The change worked when the group 2 length was too short, but it didn't work when it was too long.  Made a change to make it work when the group 2 length was too long.