Defect #1053 : Remove binary dependencies from web project.
This commit is for PART 2 of this defect. For some reason SVN failed to combine all changes into a single commit.

PART 1: (Commited previously, Rev 3714)

1) Remove Trunk/ImageServerWeb. New project resides in Trunk\ImageServer\Web


PART 2:

2) Convert existing source to ASP.NET Web Application project. This type of project allows referenced assemblies to be referenced, not copied. A web deployment project is also created for this web application to assemble the required files into a deployable folder.

A pre-build step is included to copy the plugin dlls from referenced projects to the "plugins" folder in the web app root directory. This allows developers to run the web app directly from VS for debugging purpose.

3) Refactoring: add new project (ClearCanvas.ImageServer.Web.Common for common business classes used in the web app.