Bug: #1109
Fixed a problem where we were not calculating group lengths properly for OB tags.  This had the impact that we were not calculating the group 2 length properly in a DICOM file if the tag File Meta Information Version was included in the group 2 elements (file Meta Information).  We could still read these DICOM files, but other libraries such as jdicom and offis could not read the files.