Bug: #1132 When X is clicked on a dialog box, it behaves as if OK were pressed - should behave as Cancel
Fixed.  When the user clicks the X button on a dialog, the ApplicationComponent Host automatically sets the ExitCode of the hosted component to Cancelled.  Note that this behaviour is specific to dialog box hosts.  Clicking X on a workspace or shelf is considered a "Normal" exit and therefore does not have this effect.  This creates a certain inconsistency in the design, but it was decided that this was the lesser evil, vs breaking existing API's.  Note that this fix required making the IApplicationComponent.ExitCode property publicly settable, which is not terribly desirable, but again, the lesser evil.