Requirement: #1030
Fixed Supervisor selection control on ReportEditorComponent.  It was not showing up correctly (dpi problem).  Also, added call to SupervisorSettings.Save(), as the supervisor was not being saved.  Removed the "Make default" check-box, as it no longer makes sense.  The selected Supervisor is automatically made the default whether you like it or not.