Bug: #1090
Bug fix submitted by Dan Blanchard on the forum for this defect. Added changes to the StorageScu component to handle when a few that is to be sent cannot be loaded from disk.  The DICOM file is failed and not sent and an extended failure description is set to tell the reason for the failure.  Slightly expanded the contributed patch to set extended failure descriptions for several other failures.