Bug: #1241 (UniqueConstraintValidationBroker may cause DB deadlocks)
From trac defect:

This occurs because UniqueConstraintValidationBroker uses a secondary connection/transaction to do validation. Ideally it would do the validation on the main connection/transaction, but due to certain limitations in NHibernate 1.2.0 this is not possible. Therefore a secondary connection is used, and this secondary transaction may try to read from a table that has already been written to by the main transaction. At the default RIS isolation level of ReadCommitted, this results in a deadlock.

Fix: the secondary connection now runs at a lower isolation level of ReadUncommitted, which means it will not block on reading from a table that has been written to by another transaction. The worst that can happen is that the validation broker will be over-zealous in its validation, as it may read data that has not yet been committed and will ultimately be rolled-back.  This isn't terrible, and in practice the chances of this happening are minute.