Bug: #971 Remove hard-code services URL
Part 1:
Added two settings classes: Ris.Server.WebServicesSettings and Ris.Client.WebServicesSettings.  These settings allow the server to specify the URL where it will host the services, and the client to specify the URL where the services are hosted, respectively. Both are stored in the app.config of their respective projects, rather than in the enterprise configuation store, because obviously the URL needs to be known before the enterprise config store can be contacted.