Requirement: #1270 (Allow reporting on multiple procedures)
There were a large number of errors in the reporting workflow business logic.  This commit has nothing to with #1270 really, it is just an attempt to correct some of these errors.  Chances are there are more corrections needed.

1. The Report.Parts collection cannot be modelled as a list.  From NHibernate docs:

"Please note that NHibernate does not support bidirectional one-to-many associations with an indexed collection (list, map or array) as the "many" end, you have to use a set or bag mapping."

Therefore I have re-modelled this as a sorted bag, so that we can still use the IList interface.

2. A number of changes to the state transition model for Report and ReportPart.