Bug: #1284
Introduced an Enabled bit into the ServiceLock table.  Also, changed the ServerRule and Device tables so that the former 'Active' column is now entitled 'Enabled', so that 'Enabled' is used consistently throughout the schema.  This required changes in the web gui, and in several other areas of code.  Also made updates to the new FilesystemReinventory ServiceLock row so that it sets the Enabled bit to 0 after it runs once.