Bug: #1323 details XML does not store URL of page that generated it
Fixed but not really tested.  Added a property called "capturePageUrl" to the XML blob that is stored by each details page, ie the blob records the URL of the page that was used to capture it.  If the data is subsequently reloaded, the page first checks for the existence of this property.  If the property exists and is different than the current window.location, then the page redirects itself to this URL.  This provides a form of versioning whereby anytime the data is reloaded it is always displayed by the same page that was used to capture it in the first place (assuming the URL is still valid, and it is up to the site to ensure this is always the case).