Bug: #1359 - EntityRef.Equals and == implmentation may be incorrect
After some thought and discussion, it was decided that these implementations should be modified so that version is included in the comparison by default.  Consequently a number of comparisons in the code were modified to use the version-insensitive overload.  Although a good deal of thought was put into assessing potential impacts of this change on the codebase, there is a small chance that it will introduce subtle bugs into the code, and therefore we should be on the lookout for potential regressions.