Requirement: #1391 Requirement: #1288
Reworked styles for new menu bar.  Started work on introducing styles consistently used throughout the GUI. Cleaned up the style sheets a bit.  Reworked a large amount of the GUI to make the design of each of the pages consistent.  The main changes related to this were moving the Filter panels and the Toolbar panels directly into the main panel for each page.  This is how the recently added ServerRule page was designed.