Requirement: #964 Improve config store performance
This commit contains some pre-requisite refactoring for this ticket.  These changes should be included in the 1.0 SDK release.

1. IConfigurationStore was re-named ISettingsStore.
2. ConfigurationStoreExtensionPoint was renamed SettingsStoreExtensionPoint.
3. ConfigurationStoreSettingsProvider was renamed SettingsStoreSettingsProvider.
4. LocalConfigurationStore was renamed LocalSettingsStore, and EnterpriseConfigurationStore renamed EnterpriseSettingsStore.
5. ISettingsStore.LoadSettingsValues was change to GetSettingsValues, and ISettingsStore.SaveSettingsValues changed to ISettingsStore.PutSettingsValues.
6. In addition to re-naming, the usage of this interface has changed (even though the method signatures are the same):

Previously, the expectation was that GetSettingsValues would be called twice, once with a value of null for the user, to retrieve the shared settings, and then again with the username, to retrieve the user settings.  Now, the expectation is that GetSettingsValues is called once, and it returns a dictionary that contains the combined shared settings and user settings.

Previously, the expectation was that PutSettingsValues was passed a dictionary containing settings values that differed from the defaults as defined by the settings class.  Now, PutSettingsValues will be passed a dictionary of "dirty" values, regardless of whether they differ from the defaults or not.

These changes were made to allow more flexibility in the implementation of ISettingsStore.  Essentially, the ISettingsStore implementation now has a lot more freedom (and responsibility).