Requirement: #1346
Partial commit.  Refactored brokers.  Replaced the UpdateBrokers and SelectBrokers with a single EntityBroker class.  Also, created a code generator to create the brokers and the model objects.  Replaced all the model objects with these new generated classes.  (Note that stored procedure brokers are still manually generated, although it may be possible to change this in the future.)

Note that there are still some stored procedures that duplicate the new broker code that will be removed.