Bug: #1019 Bug: #1326
Made changes to DeleteStudyStorage to address bug #1326.  Found a problem where if two studies were being deleted at the same time for a single patient, the patient record wasn't being cleaned up.  Because of the nature of the tests being done (timestamp based patient names and ids) this was much more likely to happen in the test environment.

Also, did some minor configuration updates to address #1019.  Note that some previous checkins mostly resolved this ticket.  Some changes had been done to indexes on the WorkQueue to help resolve this problem, and how delays were done within the WorkQueueProcessor were adjusted.

Also, this checkin includes a fix to a potential problem with acquiring WorkQueue items in a cluster.  If another server in the cluster got a work queue item at the same time, such that acquiring a lock on the work queue item failed, the work queue item was still being returned by the stored procedure.