Requirement: #735 Add password support
Added a setting that contains a regular expression used to validate whether a user-supplied password is good enough.  The default pattern checks that the password contains only word characters (letters, numbers, underscore) and is between 8-16 characters long.

Note: we probably want to change this setting in dev environments to allow 1 char passwords so we don't get annoyed.