Bug: #1497
Fixed problem with handling of AT VR tags, both in generating the XML file for the ImageServer and in how it was being parsed by the DICOM library.  There was a bug, first, that byte swapping was not done properly on AT VR tags when they were being read, which would cause these tags to be corrupted when sent out.  Secondly, the ToString() method for AT tags was not being overriden, and caused a decimal value to be output, whereas the SetStringValue was expecting a hexadecimal value when being set.  These were updated to be consistent.