Bug: #1499 Bug: #1473
Stats for study processing were not being recorded properly.  It turns out the TimeSpanStatistics.Add method did not work properly, and the stats that used that didn't work.  The Add method returned a new TimeSpan which must be reassigned to the value of the stat, which wasn't being done. Also made a few minor changes so study and series rules being processed are also included in the stats.  Also fixed a minor bug related to network stats being logged with incorrect ae title in the stats log.