Requirement: #1248 audit records should record changes to individual properties
Preliminary commit for this requirement.  The XML schema has not yet been finalized, but the general mechanism for capturing property-level changes is in place.

Also added a DateTimeUtils class to Common.Utilities, which right now just contains two methods: FormatISO and ParseISO, which deal with converting dates to an ISO string representation.  This representation is useful because it is completely unambiguous and therefore easy to parse.