Requirement: #967 XML Validation
Added a *special* tool extension point called ApplicationComponentMetaToolExtensionPoint.  This is for tools (typically admin tools) that operate on any ApplicationComponent, such as a tool for adjusting the input validation rules.  Only one action site is defined for this extension point, called "applicationcomponent-contextmenu".  Right-clicking anywhere on the background of an ApplicationComponentUserControl-derived class will show this menu.