Requirement: #1535, #1539, #1540 extended properties for staff and Practitioners
More work on UHN StaffEditorAdditionalInfoPageComponent.  Could still use cosmetic improvements, but those can always be done later.  Right now it has only a single field, to capture the HMRI number for a staff person.  There is some validation to make sure what is entered is actually a number.  The validation icon shows up on the wrong side, but this is captured as a more general defect (#1596).

Also, added javascript method Ris.setModified, which allows the page to inform the component that data has been modified, and improved the documentation in the ris.js file.