Bug: #1498
Cleaned up logging related to ImageServer deletes.  The original problem description was incorrect, we were logging when a study was being deleted, however, the ServiceLock process that identifies candidates for deletion was not logging clear information, and this caused confusion and this defect to be entered.  Added some better all around logging for the watermark management.

Also, cleaned up a minor compilation warning at the same time.