Requirement: #1510 Remove dependency of healthcare model on User
Interim commit.  Remove direct dependencies between Healthcare model and Enterprise.Authentication.  The Staff object is now associated to a user only via a UserName string.  Similarly, a Worklist is associated to users by a collection of username strings (this is a temporary measure until the entire worklist model can be re-thought).