Bug: #1470
Fix supplied on the forum [http://www.clearcanvas.ca/dnn/Community/Forums/tabid/69/forumid/9/postid/4873/view/topic/Default.aspx here].  Made a few changes to the patch in that the exception thrown is a bit more detailed.  Also made the ConnectTimeout be settable by clients, and added a change so that the connect timeout is copied when the structure is copied.