Requirement: #1530 Shared order details page
Made the initially selected page in the documentation workspace controlled by an application setting.  The default value for this setting is empty, meaning that the first tab page will be selected (i.e. it will no longer go to the exam page by default, because I think this is a bad idea, but if the site really wants this behaviour they can override it).