Requirement: #1530 Shared order details page
Created PerformedStepEditorPageProviderExtensionPoint, which allows custom pages to be provided in order to edit performed procedure step details.  If no extensions are found, a default editor is provided.  The default editor is just a DHTML component, so the site does not *need* to provide any extensions.  This really has very little to do with #1530 - I'm just adding the extension point as a matter of consistency.