Requirement #842: Header Retrieval Streaming Support in ImageServer
HeaderRetrieval service is added to the server to handle wcf call from the clients. The service interface provides GetStudyHeader() method which returns the stream object for the compressed header file. The client can load the compressed header file just like a normal stream. The compressed header file is generated on the fly. When the header file is updated, new compressed header will be generated when a new request is received.

A test application is created (ImageServer.TestApp) to simulate client requests.

Notes:

1. In the default configuration, the header retrieval service uses streaming as the transfer mode. This mode allows the server to stream the response instead of using buffer.
2. The size of the compressed header can be arbitrary large. The client should set its maxReceivedMessageSize to be sufficiently large to avoid "quota exceeeded" error. The simulated test client uses 50MB.


Additional changes:

- Added couple of methods into statistics classes to add new properties and sub-statistics. To enable statistics logging, the LogStatistics in ClearCanvas.ImageServer.Services.Streaming.HeaderRetrieval.Settings section should be set to "true"