Requirement #842: Header Retrieval Streaming Support in ImageServer
Update: Used either http or tcp binding (configurable) instead of named pipe. WCF disallows cross-machine communication using named pipe protocol. The type of binding can be change in the app.config under the BindingType setting of the ClearCanvas.ImageServer.Services.Shreds.StreamingServer.StreamingServer section. The default binding is "http".

The current implementation uses Windows credential for client authentication. It can be disabled by changing the security mode to "None" in the app.config. The code does not handle certificate based authentication. This will be added later.

For convenience, the test client is also changed so that binding and security mode can be selected from the UI instead of having to modify the app.config directly.