Requirement #842: Header Retrieval Streaming Support in ImageServer
Update: Added certificate based authentication to the server.

The self-signed certificate for the server can be setup using InstallServerCert.bat. By default, the configuration specifies the server to use this certificate. If needed, the instruction to change the certificate can be found in SetupCertificates.Readme.txt The self-signed certificate for test client can be installed using InstallClientCert.bat. If the client and server runs on different machines, this certificate must be imported into the server "Trusted People" store.

Other changes:
- Modified the test application to present all tests on the screen.