Requirement: #1130
Modified IDicomCodec to now include a method for decompressing a single frame.  Added a GetFrame() method to DicomCompressedPixelData to get a decompressed frame from the original image, which uses the IDicomCodec interface.  Also cleaned up the factory interface for codecs and how its used.  Added a TransferSyntax property to DicomMessageBase, and implementing setting the property for DicomMessage when we read a message off the network, so we can identify its transfer syntax.  Also added new methods for doing conversion to a dicom file when sending over the network automatically using the registered codecs if possible.